// src/app/types/types.ts
export type TProduct = {
  _id?: string;
  vegetable_name: string;
  image: string;
  description: string;
  rating: number;
  review: string;
  images: string[];
  brand: string;
  price: number;
  quantity_available_kg: number;
  discount: number;
  category: string;
  origin: string;
  organic: boolean;
  categoryName: string;
};